requirements.txt
fastapi==0.115.0
Pillow==10.3.0
pycocotools==2.0.8
torch==2.3.0
torchvision==0.18.0
uvicorn==0.30.1
